update: changed data types due to using to large data types for the range am working with

+++ src/operators/decimal_to_binary.rs
@@ -1,4 +1,6 @@
-fn decimal_to_binary(mut decimal_num: u32) -> [u32; 8] {
+// Takes a decimal number as argument and returns and array of bits with the
+// value of the decimal argument.
+fn decimal_to_binary(mut decimal_num: u8) -> [u8; 8] {
     let mut result = [0; 8];
     
     for i in (0..8).rev() {
